backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6
httpx==0.26.0
chromadb==0.4.22
langchain==0.1.0
langchain-anthropic==0.1.0
anthropic==0.8.1
